+++ crates/stdto_core/src/error.rs
@@ -1,7 +1,19 @@
 #[derive(thiserror::Error, Debug)]
 pub enum Error {
     #[error("bytes conversion error: {0}")]
-    BytesConversionError(#[from] bincode::Error),
+    Bytes(#[from] bincode::Error),
+    #[error("io error: {0}")]
+    Io(#[from] std::io::Error),
+    #[error("fmt error: {0}")]
+    Fmt(#[from] std::fmt::Error),
+    #[error("utf error: {0}")]
+    Utf8(#[from] std::str::Utf8Error),
+    #[error("parse int error: {0}")]
+    ParseInt(#[from] std::num::ParseIntError),
+    #[error("out of bounds error: {0} < {1}")]
+    OutOfBounds(usize, usize),
+    #[error("odd length")]
+    OddLength,
 }
 
 pub type Result<T> = std::result::Result<T, Error>;