+++ crates/stdto_core/src/lib.rs
@@ -1,7 +1,10 @@
+mod enums;
 mod traits;
+
+pub use enums::HexMode;
+pub use traits::{ToBytes, ToHash, ToHex};
 
 pub mod error;
 
-pub use traits::ToBytes;
-
+pub extern crate digest;
 pub extern crate serde;