+++ tests/to_hex.rs
@@ -1,0 +1,1 @@
+//! refer to the [`stdto_core`] crate, traits.rs unit test.