+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## [unreleased] - ReleaseDate
 
+## Added
+
+- Added `ToHash` and `stdto::hash` conversion.
+- Added `ToHex` and implement it to AsRef<[u8]>.
+
+## Improved
+
+- Now customizable `ToBytes` trait.
+
+Released by [@just-do-halee](https://github.com/just-do-halee).
+
 ## [0.1.0] - 2022-12-10
 
 Released by [@just-do-halee](https://github.com/just-do-halee).